requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-taggit==0.23.0
idna==2.7
Markdown==2.6.2
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.6.1
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.4.1
urllib3==1.23
view==0.1
